+++ programs/size_checker/main.cpp
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2015 Cryptonomex, Inc., and contributors.
+ * Copyright (c) 2020-2023 Revolution Populi Limited, and contributors.
  *
  * The MIT License
  *
@@ -22,13 +23,13 @@
  * THE SOFTWARE.
  */
 
-#include <graphene/protocol/block.hpp>
-#include <graphene/protocol/fee_schedule.hpp>
-
 #include <fc/io/json.hpp>
 #include <fc/io/raw.hpp>
 #include <fc/variant.hpp>
 #include <fc/variant_object.hpp>
+
+#include <graphene/protocol/block.hpp>
+#include <graphene/protocol/fee_schedule.hpp>
 
 #include <algorithm>
 #include <iostream>
@@ -85,7 +86,7 @@ int main( int argc, char** argv )
 
       idump( (witnesses) );
 
-      for( int32_t i = 0; i < op.count(); ++i )
+      for( size_t i = 0; i < op.count(); ++i )
       {
          op.set_which(i);
          op.visit( size_check_type_visitor(i) );
